requirements.txt
asgiref==3.4.1
attrs==21.4.0
Django==4.0.1
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
inflection==0.5.1
jsonschema==4.4.0
PyJWT==2.3.0
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
